% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisar-package.R
\docType{package}
\name{lisar-package}
\alias{lisar}
\alias{lisar-package}
\title{lisar: Calculate LISA differences.}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ben Hamilton \email{first.last@example.com}

}
\keyword{internal}
