% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isa.R
\name{SavingsAccount}
\alias{SavingsAccount}
\title{R6 Class Representing An ISA}
\description{
Represent the basic functionality of either a cash ISA or a cash LISA, including
government bonuses and penalties.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SavingsAccount-new}{\code{SavingsAccount$new()}}
\item \href{#method-SavingsAccount-deposit}{\code{SavingsAccount$deposit()}}
\item \href{#method-SavingsAccount-withdraw}{\code{SavingsAccount$withdraw()}}
\item \href{#method-SavingsAccount-add_interest}{\code{SavingsAccount$add_interest()}}
\item \href{#method-SavingsAccount-get_value}{\code{SavingsAccount$get_value()}}
\item \href{#method-SavingsAccount-print}{\code{SavingsAccount$print()}}
\item \href{#method-SavingsAccount-clone}{\code{SavingsAccount$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SavingsAccount-new"></a>}}
\if{latex}{\out{\hypertarget{method-SavingsAccount-new}{}}}
\subsection{Method \code{new()}}{
Create a new savings account
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsAccount$new(type = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type of account. Either "isa" or "lisa"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
New \code{SavingsAccount} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SavingsAccount-deposit"></a>}}
\if{latex}{\out{\hypertarget{method-SavingsAccount-deposit}{}}}
\subsection{Method \code{deposit()}}{
Deposit an amount of money in the account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsAccount$deposit(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Money to deposit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SavingsAccount-withdraw"></a>}}
\if{latex}{\out{\hypertarget{method-SavingsAccount-withdraw}{}}}
\subsection{Method \code{withdraw()}}{
Withdraw all funds from the account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsAccount$withdraw(house_price)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{house_price}}{Price of the house being purchased}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SavingsAccount-add_interest"></a>}}
\if{latex}{\out{\hypertarget{method-SavingsAccount-add_interest}{}}}
\subsection{Method \code{add_interest()}}{
Add interest to the account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsAccount$add_interest(rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rate}}{Interest rate (AER)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SavingsAccount-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-SavingsAccount-get_value}{}}}
\subsection{Method \code{get_value()}}{
Get the current value of the account
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsAccount$get_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SavingsAccount-print"></a>}}
\if{latex}{\out{\hypertarget{method-SavingsAccount-print}{}}}
\subsection{Method \code{print()}}{
Overwrite default printing behaviour.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsAccount$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unused}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SavingsAccount-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SavingsAccount-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SavingsAccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
